+++ Netch/Servers/Shadowsocks/Models/ShadowsocksConfig.cs
@@ -3,7 +3,7 @@ namespace Netch.Servers.Shadowsocks.Models
     public class ShadowsocksConfig
     {
         public string server { get; set; }
-        public int server_port { get; set; }
+        public ushort server_port { get; set; }
         public string password { get; set; }
         public string method { get; set; }
         public string remarks { get; set; }
